Refactor code to reduce duplication, by moving pattern matching of version incompatible expression types into version specific folders
