+++ src/main/scala/io/github/spark_redshift_community/spark/redshift/pushdown/querygeneration/MiscStatement.scala
@@ -19,15 +19,16 @@ package io.github.spark_redshift_community.spark.redshift.pushdown.querygenerati
 
 import io.github.spark_redshift_community.spark.redshift.pushdown.{ConstantString, EmptyRedshiftSQLStatement, IntVariable, RedshiftSQLStatement}
 import io.github.spark_redshift_community.spark.redshift.{RedshiftFailMessage, RedshiftPushdownUnsupportedException}
-import org.apache.spark.sql.catalyst.expressions.{Alias, Ascending, Attribute, CaseWhen, Cast, Coalesce, Descending, Expression, If, In, InSet, MakeDecimal, NullsFirst, NullsLast, ScalarSubquery, SortOrder, UnscaledValue}
+import org.apache.spark.sql.catalyst.expressions.{Alias, Ascending, Attribute, CaseWhen, Coalesce, Descending, Expression, If, In, InSet, Literal, MakeDecimal, NullsFirst, NullsLast, SortOrder, UnscaledValue}
 import org.apache.spark.sql.types._
+import org.apache.spark.unsafe.types.UTF8String
 
 /** Extractors for everything else. */
 private[querygeneration] object MiscStatement {
 
   def unapply(
-    expAttr: (Expression, Seq[Attribute])
-  ): Option[RedshiftSQLStatement] = {
+               expAttr: (Expression, Seq[Attribute])
+             ): Option[RedshiftSQLStatement] = {
     val expr = expAttr._1
     val fields = expAttr._2
 
@@ -42,7 +43,7 @@ private[querygeneration] object MiscStatement {
 
       // To support spark 3.1 as below
       // case Cast(child, t, _) =>
-      case c @ Cast(child, t, _, _) if !c.ansiEnabled =>
+      case CastExtractor(child, t, ansiEnabled) if !ansiEnabled =>
         getCastType(t) match {
           case Some(cast) =>
             // For known unsupported data conversion, raise exception to break the
@@ -83,9 +84,9 @@ private[querygeneration] object MiscStatement {
 
       case If(child, trueValue, falseValue) =>
         ConstantString("CASE WHEN") +  convertStatement(child, fields) +
-              ConstantString("THEN") + convertStatement(trueValue, fields) +
-              ConstantString("ELSE") + convertStatement(falseValue, fields) +
-              ConstantString("END")
+          ConstantString("THEN") + convertStatement(trueValue, fields) +
+          ConstantString("ELSE") + convertStatement(falseValue, fields) +
+          ConstantString("END")
 
       case In(child, list) =>
         blockStatement(
@@ -94,12 +95,12 @@ private[querygeneration] object MiscStatement {
         )
 
       case InSet(child, hset) =>
-        convertStatement(In(child, MiscStatementCommon.setToExpr(hset)), fields)
+        convertStatement(In(child, setToExpr(hset)), fields)
 
       case MakeDecimal(child, precision, scale, _) =>
         ConstantString("CAST") + blockStatement(
           blockStatement(convertStatement(child, fields) + "/ POW(10," +
-                          IntVariable(Some(scale)) + ")"
+            IntVariable(Some(scale)) + ")"
           ) + " AS DECIMAL(" + IntVariable(Some(precision)) + "," +
             IntVariable(Some(scale)) + ")"
         )
@@ -112,15 +113,14 @@ private[querygeneration] object MiscStatement {
         blockStatement(convertStatement(child, fields)) + "DESC NULLS FIRST"
       case SortOrder(child, Descending, NullsLast, _) =>
         blockStatement(convertStatement(child, fields)) + "DESC NULLS LAST"
-      
+
       // Spark 3.2 introduces below new field
       //   joinCond: Seq[Expression] = Seq.empty
       // So support to pushdown, if joinCond is empty.
       // https://github.com/apache/spark/commit/806da9d6fae403f88aac42213a58923cf6c2cb05
       // To support spark 3.1
       //      case ScalarSubquery(subquery, _, _) =>
-      // currently ignoring hintinfo which is last parameter of match
-      case ScalarSubquery(subquery, _, _, joinCond, _) if joinCond.isEmpty =>
+      case ScalarSubqueryExtractor(subquery, _, _, joinCond) if joinCond.isEmpty =>
         blockStatement(new QueryBuilder(subquery).statement)
 
       case UnscaledValue(child) =>
@@ -156,5 +156,24 @@ private[querygeneration] object MiscStatement {
 
       case _ => null
     })
+  }
+
+  final def setToExpr(set: Set[Any]): Seq[Expression] = {
+    set.map {
+      case d: Decimal => Literal(d, DecimalType(d.precision, d.scale))
+      case s @ (_: String | _: UTF8String) => Literal(s, StringType)
+      case i: Integer => Literal(i, IntegerType)
+      case d: Double => Literal(d, DoubleType)
+      case e: Expression => e
+      case default =>
+        // This exception will not break the connector. It will be caught in
+        // QueryBuilder.treeRoot.
+        throw new RedshiftPushdownUnsupportedException(
+          RedshiftFailMessage.FAIL_PUSHDOWN_SET_TO_EXPR,
+          s"${default.getClass.getSimpleName} @ MiscStatement.setToExpr",
+          "Class " + default.getClass.getName + " is not supported in the 'in()' expression",
+          false
+        )
+    }.toSeq
   }
 }
