+++ src/main/scala/io/github/spark_redshift_community/spark/redshift_spark_3_3_/pushdown/querygeneration/GetMapValueExtractor.scala
@@ -1,0 +1,10 @@
+package io.github.spark_redshift_community.spark.redshift.pushdown.querygeneration
+
+import org.apache.spark.sql.catalyst.expressions.{Expression, GetMapValue}
+
+object GetMapValueExtractor {
+  def unapply(expr: Expression): Option[(Expression, Expression, Boolean)] = expr match {
+    case GetMapValue(child, key, failOnError) => Some(child, key, failOnError)
+    case _ => None
+  }
+}
