+++ src/main/scala/io/github/spark_redshift_community/spark/redshift_spark_3_3_/pushdown/querygeneration/PromotePrecisionExtractor.scala
@@ -1,0 +1,10 @@
+package io.github.spark_redshift_community.spark.redshift.pushdown.querygeneration
+
+import org.apache.spark.sql.catalyst.expressions.{Expression, PromotePrecision}
+
+private[querygeneration] object PromotePrecisionExtractor {
+  def unapply(expr: Expression): Option[Expression] = expr match {
+    case PromotePrecision(expression) => Some(expression)
+    case _ => None
+  }
+}
