+++ src/main/scala/io/github/spark_redshift_community/spark/redshift_spark_3_3_/pushdown/querygeneration/ScalarSubqueryExtractor.scala
@@ -1,0 +1,13 @@
+package io.github.spark_redshift_community.spark.redshift.pushdown.querygeneration
+
+import org.apache.spark.sql.catalyst.expressions.{ExprId, Expression, ScalarSubquery}
+import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
+
+private[querygeneration] object ScalarSubqueryExtractor {
+  def unapply(expr: Expression): Option[(LogicalPlan, Seq[Expression], ExprId, Seq[Expression])] =
+    expr match {
+      case ScalarSubquery(plan, outerAttrs, exprId, joinCond) =>
+        Some(plan, outerAttrs, exprId, joinCond)
+      case _ => None
+  }
+}
