+++ src/main/scala/io/github/spark_redshift_community/spark/redshift_spark_3_4_/pushdown/querygeneration/CastExtractor.scala
@@ -1,0 +1,11 @@
+package io.github.spark_redshift_community.spark.redshift.pushdown.querygeneration
+
+import org.apache.spark.sql.catalyst.expressions.{Cast, Expression}
+import org.apache.spark.sql.types.DataType
+
+private[querygeneration] object CastExtractor {
+  def unapply(expr: Expression): Option[(Expression, DataType, Boolean)] = expr match {
+    case c @ Cast(child, dataType, _, _) => Some(child, dataType, c.ansiEnabled)
+    case _ => None
+  }
+}
