+++ src/main/scala/io/github/spark_redshift_community/spark/redshift_spark_3_4_/pushdown/querygeneration/PromotePrecisionExtractor.scala
@@ -1,0 +1,8 @@
+package io.github.spark_redshift_community.spark.redshift.pushdown.querygeneration
+
+import org.apache.spark.sql.catalyst.expressions.Expression
+
+private[querygeneration] object PromotePrecisionExtractor {
+  // Always return none, this operator was removed from 3.4
+  def unapply(expr: Expression): Option[Expression] = None
+}
